Added feature to produce the amplitude waterfall plots by default whenever gen_caltables is called. Additionally now the code supports conditions where all msfiles are present, or when calibration is not successfull for some bands.
New code has also been put in to help choosing appropriate time of calibration. This
becomes needed as now we are always in daytime settings and hence we can choose good
times for calibration, when the sources are high up in sky. While the functionality
for calibrating the data still needs to be input manually, the helper function will
allow the user to choose appropriate time.